apiGrupo2/apps/pedidos/serializers.py
# Aca construirmos los serializers para el modelo de Pedidos
from rest_framework import serializers
from .models import Usuario, Pedido, Producto, DetallePedido, Entrega, Reporte, Notificacion

# --- Serializers Base ---

class UsuarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Usuario
        # Exponemos solo estos campos (role lo definimos dentro de models.py los demas vienen por defecto con AbstractUser)
        fields = ['id', 'username', 'email', 'role', 'password']
        extra_kwargs = {'password': {'write_only': True}}

class ProductoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Producto
        fields = '__all__'

# --- Serializers Dependientes ---

class PedidoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Pedido
        fields = '__all__'
        
    def get_entrega(self, obj):
        entrega = getattr(obj, 'entrega', None)
        if entrega:
            return {
                "id": entrega.id,
                "estado": entrega.estado,
                "repartidor": entrega.repartidor.username,
                "fecha_asignacion": entrega.fecha_asignacion,
            }
        return None

class DetallePedidoSerializer(serializers.ModelSerializer):
    class Meta:
        model = DetallePedido
        fields = '__all__'

class EntregaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Entrega
        fields = '__all__'

class ReporteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reporte
        fields = '__all__'

class NotificacionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notificacion
        fields = '__all__'